7039 Recombine CDK deployment test
